% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DisclosureSettings.FredaP21}
\alias{DisclosureSettings.FredaP21}
\title{Disclosure settings for FredaP21}
\format{
\subsection{\code{DisclosureSettings.FredaP21}}{

\code{list}
\describe{
\item{Profile}{Can be 'strict' or 'loose'}
\item{NThreshold}{The minimum sample size required for transmission of aggregated data to client}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsP21Curator/blob/main/Development/MetaData}
}
\usage{
DisclosureSettings.FredaP21
}
\description{
A \code{list} containing settings concerning data privacy and disclosure mitigation
}
\author{
Bastian Reiter
}
\keyword{datasets}
