% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_Departments}
\alias{Meta_Departments}
\title{Meta data on hospital department codes}
\format{
\subsection{\code{Meta_Departments}}{

\code{tibble}
\describe{
\item{DepartmentCode}{}
\item{DepartmentOriginalLabel}{}
\item{Department}{}
\item{OperatingSpecialty}{}
\item{Subspecialty}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsP21Curator/blob/main/Development/MetaData}
}
\usage{
Meta_Departments
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
