% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_DischargeReasons}
\alias{Meta_DischargeReasons}
\title{Meta data on discharge reason coding}
\format{
\subsection{\code{Meta_DischargeReasons}}{

\code{tibble}
\describe{
\item{DischargeReasonCode}{}
\item{DischargeReasonOriginalLabel}{}
\item{DischargeCategory}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsP21Curator/blob/main/Development/MetaData}
}
\usage{
Meta_DischargeReasons
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
