% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P21.AssessComorbidityDS.R
\name{P21.AssessComorbidityDS}
\alias{P21.AssessComorbidityDS}
\title{P21.AssessComorbidityDS}
\usage{
P21.AssessComorbidityDS(
  DiagnosisData.S,
  DiagnosticCodeFeature.S,
  IDFeature.S,
  IgnoredCategories.S = NULL,
  ReturnScoreValueOnly.S = FALSE,
  ScoreFeatureName.S = "ComorbidityScore",
  Arg.map.S = "charlson_icd10_quan",
  Arg.assign0.S = TRUE,
  Arg.weights.S = "quan"
)
}
\arguments{
\item{DiagnosisData.S}{\code{string} - Name of the \code{data.frame} holding diagnostic code data and a grouping identifier feature}

\item{DiagnosticCodeFeature.S}{\code{string} - Name of the feature containing diagnostic codes}

\item{IDFeature.S}{\code{string} - Name of the feature containing patient- or case-specific identifier (or any other feature grouping the diagnostic codes)}

\item{IgnoredCategories.S}{Optional \code{character vector} - Containing morbidity categories that should be ignored in score calculation. See documentation on \code{comorbidity::comorbidity()} for reference.}

\item{ReturnScoreValueOnly.S}{\code{logical} - Whether the returned \code{data.frame} should only contain the feature with comorbidity score values and not the category-specific features}

\item{ScoreFeatureName.S}{\code{string} - The name of the feature that holds the comorbidity score values in the returned \code{data.frame}}

\item{Arg.map.S}{\code{string} - The value for argument 'map' in \code{comorbidity::comorbidity()}}

\item{Arg.assign0.S}{\code{logical} - The value for argument 'assign0' in \code{comorbidity::comorbidity()}}

\item{Arg.weights.S}{\code{string} - The value for argument 'weights' in \code{comorbidity::score()}}
}
\value{
A \code{data.frame}
}
\description{
Using functionality in package \code{comorbidity}, map ICD codes to morbidity categories and calculate a comorbidity score.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
