% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P21.AugmentDataDS.R
\name{P21.AugmentDataDS}
\alias{P21.AugmentDataDS}
\title{P21.AugmentDataDS}
\usage{
P21.AugmentDataDS(
  CuratedDataSetName.S = "P21.CuratedDataSet",
  Settings.S = list(CreateSubsets = list(Cancer = TRUE, HIVCancer = FALSE))
)
}
\arguments{
\item{CuratedDataSetName.S}{\code{character} - Name of the Curated Data Set object on server - Default: 'P21.CuratedDataSet'}

\item{Settings.S}{\code{list} - Settings passed to function
\itemize{\item EventFeatures \code{list}
\itemize{\item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_EventFeatures}
\item Profile \code{character} - Profile name defining rule set to be used for event feature engineering. Profile name must be stated in \code{EventFeatures$RuleSet} - Default: 'Default'}}}
}
\value{
A \code{list} containing the following objects:
\itemize{\item AugmentedDataSet \code{list}
\itemize{\item Events
\item DiseaseCourse
\item Therapy
\item Diagnosis
\item Patient}
\item AugmentationReport \code{list}
\item AugmentationMessages \code{list}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Transforms Curated Data Set (CDS) into Augmented Data Set (ADS)
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
