% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P21.DrawSampleDS.R
\name{P21.DrawSampleDS}
\alias{P21.DrawSampleDS}
\title{P21.DrawSampleDS}
\usage{
P21.DrawSampleDS(RawDataSetName.S = "P21.RawDataSet", SampleSize.S = 100)
}
\arguments{
\item{RawDataSetName.S}{\code{string} - Name of Raw Data Set object (list) on server - Default: 'P21.RawDataSet'}

\item{SampleSize.S}{\code{count} - Number of patients in sample}
}
\value{
A \code{list} containing a subset of Raw Data Set
}
\description{
Draws a sample (subset) from Raw Data Set
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
