% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P21.LinkToCCP.R
\name{P21.LinkToCCPDS}
\alias{P21.LinkToCCPDS}
\title{P21.LinkToCCPDS}
\usage{
P21.LinkToCCPDS(
  P21CDSName.S = "P21.CuratedDataSet",
  CCPADSName.S = "CCP.AugmentedDataSet",
  Tolerance.DischargeToDiagnosis.S = 30,
  Tolerance.DiagnosisToAdmission.S = 30
)
}
\arguments{
\item{P21CDSName.S}{\code{string} - Name of P21 Curated Data Set on servers}

\item{CCPADSName.S}{\code{string} - Name of CCP Augmented Data Set on servers}

\item{Tolerance.DischargeToDiagnosis.S}{\code{integer}}

\item{Tolerance.DiagnosisToAdmission.S}{\code{integer}}
}
\value{
A \code{data.frame}
}
\description{
Map P21 cases to DiagnosisIDs in CCP data
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
