% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P21.PrepareRawDataSetDS.R
\name{P21.PrepareRawDataSetDS}
\alias{P21.PrepareRawDataSetDS}
\title{P21.PrepareRawDataSetDS}
\usage{
P21.PrepareRawDataSetDS(
  RawDataSetName.S = "P21.RawDataSet",
  FeatureNameDictionary.S = list()
)
}
\arguments{
\item{RawDataSetName.S}{\code{string} - Name of Raw Data Set object (list) on server - Default: 'P21.RawDataSet'}
}
\value{
A \code{list} containing a subset of Raw Data Set
}
\description{
Load
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
