% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRawDataSetDS.R
\name{PrepareRawDataSetDS}
\alias{PrepareRawDataSetDS}
\title{PrepareRawDataSetDS}
\usage{
PrepareRawDataSetDS(RawDataSetName.S = "RawDataSet")
}
\arguments{
\item{RawDataSetName.S}{\code{string} - Name of Raw Data Set object (list) on server - Default: 'RawDataSet'}
}
\value{
A \code{list} containing a subset of Raw Data Set
}
\description{
Load
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
