% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.Privacy}
\alias{Set.Privacy}
\title{Privacy settings for FredaP21}
\format{
\subsection{\code{Set.Privacy}}{

\code{list}
\describe{
\item{Profile}{Can be 'strict' or 'loose'}
\item{NThreshold}{The minimum sample size required for transmission of aggregated data to client}
}
}
}
\usage{
Set.Privacy
}
\description{
A \code{list} containing settings concerning data privacy and disclosure mitigation
}
\keyword{datasets}
