% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsFredaP21-package.R
\docType{package}
\name{dsFredaP21-package}
\alias{dsFredaP21}
\alias{dsFredaP21-package}
\title{dsFredaP21: Federated Processing of P21 data}
\description{
Server-side DataSHIELD R package to process P21 data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ccphos.netlify.app/}
  \item \url{https://github.com/BastianReiter/dsFredaP21}
}

}
\author{
\strong{Maintainer}: Bastian Reiter \email{b.reiter@med.uni-frankfurt.de} (\href{https://orcid.org/0000-0002-1088-0502}{ORCID})

}
\keyword{internal}
