% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P21.AugmentDataDS.R
\name{P21.AugmentDataDS}
\alias{P21.AugmentDataDS}
\title{P21.AugmentDataDS}
\usage{
P21.AugmentDataDS(
  CuratedDataSetName.S = "P21.CuratedDataSet",
  Settings.S = list(CreateSubsets = list(Cancer = TRUE, HIVCancer = FALSE))
)
}
\arguments{
\item{CuratedDataSetName.S}{\code{character} - Name of the Curated Data Set object on server - Default: 'P21.CuratedDataSet'}

\item{Settings.S}{\code{list} - Settings passed to function
\itemize{\item CreateSubsets \code{list}
\itemize{\item Cancer \code{logical} - Default: \code{TRUE}
\item HIVCancer \code{logical} - Default: \code{FALSE}}}}
}
\value{
A \code{list} containing the following objects:
\itemize{\item AugmentedDataSet \code{list}
\itemize{\item Case \code{data.frame}
\item Diagnosis \code{data.frame}
\item Events \code{data.frame}
\item Patient \code{data.frame}
\item PatientCancer \code{data.frame}
\item PatientHIVCancer \code{data.frame}
\item Procedures \code{data.frame} }
\item AugmentationReport \code{list}
\item Messages \code{list}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Transforms Curated Data Set (CDS) into Augmented Data Set (ADS)
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
