% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P21.PrepareRawDataSetDS.R
\name{P21.PrepareRawDataSetDS}
\alias{P21.PrepareRawDataSetDS}
\title{P21.PrepareRawDataSetDS}
\usage{
P21.PrepareRawDataSetDS(
  RawDataSetName.S = "P21.RawDataSet",
  FeatureNameDictionary.S = list()
)
}
\arguments{
\item{RawDataSetName.S}{\code{string} - Name of Raw Data Set object (list) on server - Default: 'P21.RawDataSet'}

\item{FeatureNameDictionary.S}{Optional \code{list} containing dictionary data for feature name harmonization (Form: \code{list(Department = c(FAB = "Fachabteilung"))})}
}
\value{
A \code{list} containing
\itemize{ \item The input RawDataSet (\code{list}) with harmonized feature names
\item Messages (\code{character vector}) }
}
\description{
Try to harmonize feature names in RDS tables (prior to data curation) employing Fuzzy String Matching and Dictionary look-up.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
